test: split user action tests
